Move AggregationSearcher into a separate file
